refactor: Relocate backend types package
This commit moves the backendtypes package to a more general location under the lib directory. This is purely a structural refactoring with no logic changes. All import paths have been updated accordingly.

Split up of #1706
