+++ backend/pkg/httpserver/list_chromium_usage.go
@@ -19,7 +19,7 @@ import (
 	"errors"
 	"log/slog"
 
-	"github.com/GoogleChrome/webstatus.dev/lib/gcpspanner/spanneradapters/backendtypes"
+	"github.com/GoogleChrome/webstatus.dev/lib/backendtypes"
 	"github.com/GoogleChrome/webstatus.dev/lib/gen/openapi/backend"
 )
