+++ backend/pkg/httpserver/remove_user_saved_search_bookmark_test.go
@@ -20,7 +20,7 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/GoogleChrome/webstatus.dev/lib/gcpspanner/spanneradapters/backendtypes"
+	"github.com/GoogleChrome/webstatus.dev/lib/backendtypes"
 )
 
 func TestRemoveUserSavedSearchBookmark(t *testing.T) {
